last update using libraries etc

+++ ecommerce.js
@@ -1,27 +1,49 @@
+//Logre implementar un api local con fetch y promesas, sin embargo a la hora de querer realizar el metodo del carrito por mas que intente no pude lograr filtrar por id!
+
+
 class Product {
-    constructor(id, brand, model, price, amount) {
+    constructor(id, brand, model, price, img) {
         this.id = id;
         this.brand = brand;
         this.model = model;
         this.price = price;
+        this.img = img;
         this.amount = 1;
     }
 }
 
 const products = [];
-const product1 = new Product(1, "Nike", "Sportswear Blue T-Shirt", 40, 9);
+
+const product1 = new Product(1, "GVNG", "Anorak (Lightblue and Pink)", 140, "img/gvngswanorak.png");
 products.push(product1);
-const product2 = new Product(2, "Adidas", "Classic Green Jacket", 80, 1);
+const product2 = new Product(2, "GVNG", "Anorak (White and Orange)", 130, "img/gvngswanorak2.png");
 products.push(product2);
-const product3 = new Product(3, "Vans", "Old-School Red Trainers", 70, 16);
+const product3 = new Product(3, "BASTARD", "Hoodie (Black)", 150, "img/hoodiebastard.png");
 products.push(product3);
-const product4 = new Product(4, "Puma", "New Wave Orange T-shirt", 30, 10);
+const product4 = new Product(4, "TYLER", "Hoodie Tyler The Creator (Black)", 190, "img/hoodietyler.png");
 products.push(product4);
+const product5 = new Product(5, "Levis", "Jeans Unisex Summer (Blue)", 140, "img/pantsblue.png");
+products.push(product5);
+const product6 = new Product(6, "Levis", "Jeans Unisex Summer (Grey)", 120, "img/pantsummer.png");
+products.push(product6);
+const product7 = new Product(7, "2pac", "T-Shirt 2pac All Eyes On Me", 80, "img/remera2pac.png");
+products.push(product7);
+const product8 = new Product(8, "Cross Clothing", "T-Shirt Cross (White)", 70, "img/remeracross.png");
+products.push(product8);
+const product9 = new Product(9, "Notorious", "T-shirt BIGGIE", 80, "img/remeranotorious.png");
+products.push(product9);
+const product10 = new Product(10, "A$AP", "T-SHIRT A$AP ROCKY ", 60, "img/remerarocky.png");
+products.push(product10);
+const product11 = new Product(11, "BASTARD", "Shorts Bastard (Black)", 90, "img/shortbastard.png");
+products.push(product11);
+const product12 = new Product(12, "Loop", "Shorts LoopCompany (Black)", 100, "img/shortloopcompany.png");
+products.push(product12);
 
 let cart = []
-if(localStorage.getItem("cart")) {
+if (localStorage.getItem("cart")) {
     cart = JSON.parse(localStorage.getItem("cart"));
 };
+
 
 const divProducts = document.getElementById("divProducts");
 
@@ -30,14 +52,12 @@ const divProducts = document.getElementById("divProducts");
 const showProducts = () => {
     products.forEach(element => {
         const card = document.createElement("div");
-        card.classList.add("col-xl-3", "col-md-6", "col-xs-12");
         card.innerHTML = `
-        <div class="card">
-            <div class="card-body">
-            <h5 class="card-title">${element.brand}, ${element.model}</h5>
-            <p class="card-text"> $${element.price}</p>
-            <button class="btn btn-primary" id="button${element.id}"> Add to cart</button>
-            </div>
+        <div class="producto">
+            <img class="producto__imagen" src="${element.img}" alt="imagen camisa"> 
+            <p class="producto__nombre">${element.brand}, ${element.model}</h5>
+            <p class="producto__precio"> $${element.price}</p>
+            <button class="botonCart" id="button${element.id}"> Add to cart</button>
         </div>
         `
 
@@ -47,6 +67,20 @@ const showProducts = () => {
         const button = document.getElementById(`button${element.id}`);
         button.addEventListener("click", () => {
             addToCart(element.id)
+        })
+
+        button.addEventListener("click", () => {
+            Toastify({
+                text: "Product added to cart",
+                className: "Toastify__toast",
+                duration: 3000,
+                gravity: "bottom",
+                position: "right",
+                style: {
+                    background: "linear-gradient(to right, #0a1425, #0a1425)"
+                },
+
+            }).showToast();
         })
     });
 }
@@ -60,10 +94,11 @@ const addToCart = (id) => {
         productCart.amount++;
     } else {
         cart.push(product);
-        localStorage.setItem("cart",JSON.stringify(cart));
+        localStorage.setItem("cart", JSON.stringify(cart));
     }
     calculateTotal();
 }
+
 
 // Cart
 
@@ -77,14 +112,14 @@ const showCart = () => {
     divCart.innerHTML = "";
     cart.forEach(element => {
         const card = document.createElement("div");
-        card.classList.add("col-xl-3", "col-md-6", "col-xs-12");
         card.innerHTML = `
-        <div class="card">
-            <div class="card-body">
-            <h5 class="card-title">${element.brand}, ${element.model}</h5>
-            <p class="card-text"> $${element.price}</p>
-            <p class="card-text"> ${element.amount}</p>
-            <button class="btn btn-primary" id="delete${element.id}"> Delete product</button>
+        <div class="grid">
+            <div class="producto">
+            <img class="producto__imagen" src="${element.img}" alt="imagen camisa">
+            <p class="producto__nombre">${element.brand}, ${element.model}</h5>
+            <p class="producto__precio"> $${element.price}</p>
+            <p class=""> ${element.amount}</p>
+            <button class="botonCart" id="delete${element.id}"> Delete product</button>
             </div>
         </div>
         `
@@ -104,13 +139,32 @@ const deleteFromCart = (id) => {
     const index = cart.indexOf(product);
     cart.splice(index, 1);
     showCart();
-    localStorage.setItem("cart", JSON.stringify(cart)); 
+    localStorage.setItem("cart", JSON.stringify(cart));
 }
 
 const clearCart = document.getElementById("clearCartButton");
 clearCart.addEventListener("click", () => {
-    deleteAllProducts();
+    Swal.fire ({
+        title:"Are you sure you want to delete everything?",
+        icon:"warning",
+        confirmButtonText: "Accept",
+        showCancelButton: true,
+        cancelButtonText:"Cancel",
+        cancelButtonColor: "#FFCE00",
+        confirmButtonColor:  "B7950b",
+    }).then((result) => {
+        if(result.isConfirmed) {
+            deleteAllProducts();
+            Swal.fire ( {
+                title: "Products Deleted",
+                icon: "success",
+                confirmButtonText: "Accept",
+                confirmButtonColor: "#B7950B",
+            })
+        }
+    })
 })
+
 
 
 // Method that deletes all products inside the cart.
@@ -130,5 +184,6 @@ const calculateTotal = () => {
     cart.forEach((element) => {
         totalPayment += element.price * element.amount;
     })
-    total.innerHTML = `TOTAL: $${totalPayment}`;
+    total.innerHTML = `$${totalPayment}`;
 }
+
